#882: Share code between Portable Server build and SCons build.
