+++ earth_enterprise/src/lib/python/opengee/__init__.py
@@ -1,0 +1,5 @@
+"""
+A collection of Python modules used by Open GEE during the build process.
+"""
+
+__all__ = ['c_compiler', 'environ', 'version']
